requirements.txt
django
djangorestframework
django-cors-headers
pandas
python-docx
openpyxl
gunicorn
